Did 5a,5b
Added periodic BC + made total momentum = 0

+++ Programs/atom.h
@@ -5,7 +5,7 @@
 class Atom
 {
 private:
-    float m_mass;
+    double m_mass;
 public:
     vec3 position;
     vec3 velocity;
